+++ packages/nft-checkout/src/operations/evm/evm-operation.ts
@@ -7,6 +7,7 @@ import {
   type BridgeChain,
   ChainTypes,
   DestinationTransactionStatus,
+  type HexString,
   isString,
   type TransactionBundle,
 } from '@rarimo/shared'
@@ -15,8 +16,9 @@ import {
   createSwapper,
   type IntermediateTokenOpts,
 } from '@rarimo/swap'
-import type { providers } from 'ethers'
+import { type providers, utils } from 'ethers'
 
+import { checkoutApi } from '@/api'
 import { USDC_MAP } from '@/const'
 import { CheckoutOperationStatus, OperationEventBusEvents } from '@/enums'
 import { errors } from '@/errors'
@@ -59,6 +61,11 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
   const isInitialized = ref(false)
   const status = ref(CheckoutOperationStatus.Created)
 
+  const relayer = {
+    [ChainTypes.EVM]: '0x1bbcc1c328e47805f050cb8c4bee9a6043997118',
+    [ChainTypes.Solana]: '',
+    [ChainTypes.Near]: '',
+  }
   let params = {} as CheckoutOperationParams
   let isSameChain = false
   let isMultiplePayment = false
@@ -97,6 +104,15 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
 
     swapToToken = _getSwapToToken()
     swapAmountOut = await _getSwapAmountOut()
+
+    if (params?.relayer?.[ChainTypes.EVM])
+      relayer[ChainTypes.EVM] = params.relayer[ChainTypes.EVM]
+
+    if (params?.relayer?.[ChainTypes.Solana])
+      relayer[ChainTypes.Solana] = params.relayer[ChainTypes.Solana]
+
+    if (params?.relayer?.[ChainTypes.Near])
+      relayer[ChainTypes.Near] = params.relayer[ChainTypes.Near]
 
     _emitEvent(OperationEventBusEvents.Initiated)
     _setStatus(CheckoutOperationStatus.Initialized)
@@ -172,6 +188,8 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
     }
 
     _setStatus(CheckoutOperationStatus.CheckoutStarted)
+
+    if (bundle?.salt) bundle.salt = _getSalt(bundle.salt)
 
     const result = await _checkout({
       swapper,
@@ -320,6 +338,27 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
     })
   }
 
+  const _getSalt = (salt: HexString): HexString => {
+    const sender = _getWithdrawTxSender()
+
+    return new utils.AbiCoder().encode(['bytes32', 'address'], [salt, sender])
+  }
+
+  const _getWithdrawTxSender = () => {
+    return isSameChain ? provider.address : relayer[ChainTypes.EVM]
+  }
+
+  const getBundlerAddress = async (salt: HexString) => {
+    if (!isInitialized.value) throw new errors.OperatorNotInitializedError()
+    if (!salt) throw new TypeError('Salt is required')
+
+    return checkoutApi.get('/v1/bridge/bundler-address', {
+      query: {
+        salt: _getSalt(salt),
+      },
+    })
+  }
+
   return toRaw(
     extend(
       {
@@ -331,6 +370,7 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
         getSupportedChains,
         getPaymentTokens,
         getDestinationTx,
+        getBundlerAddress,
         checkout: checkout as CheckoutOperation['checkout'],
         estimatePrice: estimatePrice as CheckoutOperation['estimatePrice'],
       },