+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-rc.25] - 2023-09-19
+### Added
+- `@rarimo/nft-checkout` - `getBundlerAddress` method
+
+### Changed
+- `@rarimo/nft-checkout` - `EVMOperation` modifying salt
+
 ## [2.0.0-rc.24] - 2023-09-12
 ### Fixed
 - `@rarimo/zkp-gen-iden3` - `endianSwappedCoreStateHashHex`
@@ -469,7 +476,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.24...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.25...HEAD
+[2.0.0-rc.25]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.24...2.0.0-rc.25
 [2.0.0-rc.24]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.23...2.0.0-rc.24
 [2.0.0-rc.23]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.22...2.0.0-rc.23
 [2.0.0-rc.22]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.21...2.0.0-rc.22